% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_bn.R
\name{custom_bn}
\alias{custom_bn}
\title{Construct a bn.fit object}
\usage{
custom_bn(dag, dist, use_bnlearn = TRUE)
}
\arguments{
\item{dag}{adjacency matrix}

\item{dist}{list of local distibution}

\item{use_bnlearn}{use \code{\link[bnlearn:bn.fit]{bnlearn::custom.fit()}} to construct the bn.fit object (implies a set of checks of the input) or hardcode such an object (with no checks).}
}
\value{
an object of class \link[bnlearn:bn.fit]{bnlearn::bn.fit}
}
\description{
Construct a \link[bnlearn:bn.fit]{bnlearn::bn.fit} object from a DAG and a distribution (set of local distributions).
}
