% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_partition.R
\name{rand_partition_tree}
\alias{rand_partition_tree}
\title{grow a tree that partition the space of \code{length(dims)} variables}
\usage{
rand_partition_tree(
  dims,
  splitprob,
  regular,
  nextsplitprob = function(x) x,
  mindepth = 1,
  maxdepth = length(dims)
)
}
\arguments{
\item{splitprob}{(numeric constant) probability of splitting a node}

\item{nextsplitprob}{(function) manipulate \code{splitprob} for each new split}

\item{doMerge}{(logical constant) randomly merge leaves in tree}
}
\description{
grow a tree that partition the space of \code{length(dims)} variables
}
