% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_tree.R
\name{rand_tree}
\alias{rand_tree}
\title{Grow a random decision tree}
\usage{
rand_tree(dims, p, branch = "", mindepth = 1, maxdepth = length(dims))
}
\arguments{
\item{dims}{}

\item{p}{}

\item{branch}{}

\item{mindepth}{}

\item{maxdepth}{}
}
\description{
Grow a random decision tree
}
\examples{
dims <- 2:4
names(dims) <- letters[seq_along(dims)]
tree <- rand_tree(dims, .5)
tree
 
tree <- rand_tree(dims, 1) 
summary(tree)$nparts == prod(dims)

tree <- rand_tree(dims, 0, mindepth = 0) 
summary.tree(tree)$nparts == 1

# mindepth and maxdepth overrules `p`
tree <- rand_tree(dims, 0, mindepth = length(dims)) 
summary.tree(tree)$nparts == prod(dims)
tree <- rand_tree(dims, 1, maxdepth = 0) 
summary.tree(tree)$nparts == 1
 
# predict   
newdata <- expand.grid(lapply(dims-1, seq.int, from = 0))
predict(tree, newdata)

# regular
tree <- rand_tree(dims, 0, mindepth = 1) 
make_regular(tree, dims)

}
