% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_cpt.R
\name{rand_tree_cpt}
\alias{rand_tree_cpt}
\title{Title}
\usage{
rand_tree_cpt(
  dims,
  p,
  mindepth = 0,
  maxdepth = length(dims),
  regular = TRUE,
  ...
)
}
\arguments{
\item{dims}{}

\item{p}{}

\item{mindepth}{}

\item{maxdepth}{}

\item{regular}{}

\item{...}{}
}
\description{
Title
}
\examples{

dims <- c(a = 2, b = 2, c = 2)
p <- 0
mindepth <- 0 
maxdepth <- 0
regular <- TRUE
rand_tree_cpt(dims, p, mindepth, maxdepth, regular, method = "constant-mean")
}
