% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_cpt.R
\name{rand_cpt}
\alias{rand_cpt}
\title{Draw a conditional probability table (CPT)}
\usage{
rand_cpt(nlev, alpha = 1, method = c("direct", "cm"), scope = character(), ...)
}
\arguments{
\item{nlev}{(integer vector) of cardinality of the outcome variable and the
parent variables (the dimensions of the CPT). First element correspond to the outcome,
the rest to its parents.}

\item{alpha}{(numeric vector)}

\item{method}{(character) Which method to use. The direct method (default) is
to assume `alpha` specifies the Dirichlet hyperparameters associated with each
outcome. Specifically, a symmetric Dirichlet distribution if `alpha` is a constant,
a single Dirichlet distribution if `alpha` is a vector of length `nlev[1]` or
a (possibly) unique Dirichlet distribution for each parent level if `alpha` is
of length `prod(nlev)`. If method equals `CM`, each conditional distribution is
drawn by perturbing a vector of means as described in XX.  `alpha` is then taken
to be the imaginary sample size / precision parameter controlling the variance of
the Dirichlet distribution.}

\item{scope}{(character vector) names of the variables. Defaults to `names(nlev)`.}
}
\value{
an array with dimension `nlev` representing a CPT.
}
\description{
Construct a random conditional probability table (CPT) by sampling independently
a vector of probabilities from a Dirichlet distribution for each parent configuration.
If the scope of the CPT is specified, the function returns an array with
named dimensions, a format accepted for local distributions by [bnlearn::bn.fit.dnode].
}
\examples{

# no parents
cpt <- rand_cpt(nlev = 2, scope = "Y")
cpt
dimnames(cpt)

# two parents
cpt <- rand_cpt(2:4, scope = c("Y", "X", "Z"))
cpt

# with local structure
\dontrun{
cpt <- rand_cpt(nlev, local_struct = "tree", prob = 1, maxdepth = .5)
}
}
\keyword{internal}
