% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_dist.R
\name{rand_dist}
\alias{rand_dist}
\alias{rand_dist_cat}
\title{Draw a random distribution over a DAG}
\usage{
rand_dist(dag, type = "cat", ...)

rand_dist_cat(
  dag,
  nlev = lengths(levels),
  levels = NULL,
  rand_cpt = randbn::rand_cpt,
  ...
)
}
\arguments{
\item{dag}{(integer matrix) an adjacency matrix}

\item{type}{(character) type of distribution.}

\item{...}{additional arguments for the function producing each local distribution.
See \code{\link[=rand_cpt]{rand_cpt()}} for categorical distributions.}

\item{nlev, levels}{specifies the cardinality of each variable, as a vector or
a list with outcomes for each node, respectively. Used to set dimension and
dimmension names for each CPT, aligned with the requirements for bn.fit.dnode objects.
If \code{levels = NULL}, the levels of node j is sat to \verb{0, 1, ..., nlev[j]-1}.
Both arguments can not be \code{NULL}.}

\item{rand_cpt}{(function) a function that generates a CPT.
The first argument must be the dimension of the CPT, derived from \code{nlev}, i.e.
\code{nlev[c(j, parentnodes)]} for node \code{j} with parents \code{parentnodes}.}
}
\value{
a list with distributions
}
\description{
Draw a random distribution over a DAG.
The function returns a list of CPTs (if `type = "categorical") stored
in a format matching \link[bnlearn:bn.fit]{bnlearn::bn.fit} class and can be used to construct
such objects.
}
\examples{

dag  <- matrix(0, 3, 3, dimnames = list(c("Z", "X", "Y"), c("Z", "X", "Y")))
dag[upper.tri(dag)] <- 1
rand_dist(dag, "cat", nlev = rep(3, 3), alpha = 1)
rand_dist(dag, "cat", nlev = rep(3, 3), rand_cpt = rand_cpt_cm, ess = 1000)
}
\seealso{
\code{\link[=rand_cpt]{rand_cpt()}}
}
