% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_partition.R
\name{rand_partition}
\alias{rand_partition}
\title{Draw a random partition of an outcome space}
\usage{
rand_partition(dims, prob, method = "tree", regular = FALSE, ...)
}
\arguments{
\item{dims}{cardinality of the variables.}

\item{prob}{parameter controlling the size of the partition.}

\item{method}{name of partitioning algorithm.}

\item{regular}{If \code{TRUE}, the partition is forced to be regular. See \link[bida:make_regular]{bida::make_regular}.}
}
\value{
an vector of length \code{prod(dims)} assigning each joint outcome to a
subset of the partition.
}
\description{
Draw a random partition of an outcome space
}
\examples{

set.seed(1)
dims <- c(2, 2, 2)
rand_partition(dims, .75, method = "tree")

# tree of given depth
rand_partition(dims, 1, "tree", maxdepth = 1)
rand_partition(dims, 1, "tree", maxdepth = .5)

# tree of given depth, forced to be regular
rand_partition(dims, 1, "tree", maxdepth = 1, regular = TRUE)
}
