% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_tree.R
\name{rand_tree}
\alias{rand_tree}
\alias{make_tree_regular}
\alias{partition_from_tree}
\title{Grow a random decision tree}
\usage{
rand_tree(dims, p, mindepth = 0, maxdepth = length(dims), regular = TRUE)

make_tree_regular(tree, dims)

partition_from_tree(tree, newdata)
}
\arguments{
\item{dims}{}

\item{p}{}

\item{mindepth}{}

\item{maxdepth}{}

\item{branch}{}
}
\description{
Grow a random decision tree
}
\examples{
dims <- 2:4
names(dims) <- letters[seq_along(dims)]
tree <- rand_tree(dims, .5)
tree

newdata <- expand.grid(lapply(dims-1, seq.int, from = 0))
partition_from_tree(tree, newdata)
 
# full depth
tree <- rand_tree(dims, 1) 
summary(tree)$nparts == prod(dims)

# no splits
tree <- rand_tree(dims, 0) 
summary.tree(tree)$nparts == 1

# stump / one split
tree <- rand_tree(dims, 0, mindepth = 1)
tree 

# mindepth and maxdepth overrules `p`
tree <- rand_tree(dims, 0, mindepth = length(dims)) 
summary.tree(tree)$nparts == prod(dims)
tree <- rand_tree(dims, 1, maxdepth = 0) 
summary.tree(tree)$nparts == 1
 

# regular
tree <- rand_tree(dims, 0, mindepth = 1) 
tree <- make_regular(tree, dims)

}
