% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDirichlet.R
\name{rDirichlet}
\alias{rDirichlet}
\title{Wrapper around rBeta2009::rdirichlet that always return an n-by-k matrix}
\usage{
rDirichlet(n, alpha, k = length(alpha))
}
\description{
Wrapper around rBeta2009::rdirichlet that always return an n-by-k matrix
}
